Change structure of python module (avoids issues with potential ROS custom messages)
